+++ app/components/ui/badge.tsx
@@ -1,12 +1,6 @@
+import { Badge, badgeVariants } from "@chester-hill-solutions/shad-cc/badge";
 import type { ComponentProps } from "react";
 import type { VariantProps } from "class-variance-authority";
 
-import {
-  Badge,
-  badgeVariants,
-} from "@chester-hill-solutions/shad-cc/badge";
-
-export type BadgeProps = ComponentProps<"span"> &
-  VariantProps<typeof badgeVariants>;
-
+export type BadgeProps = ComponentProps<"span"> & VariantProps<typeof badgeVariants>;
 export { Badge, badgeVariants };
