+++ app/lib/messaging-onboarding/goals.ts
@@ -5,10 +5,9 @@
 
 import {
   WORKSPACE_ONBOARDING_GOAL_VALUES,
-  type WorkspaceOnboardingChannel,
   type WorkspaceOnboardingGoal,
-  type WorkspaceOperatingCountry,
-} from "@/lib/types";
+} from "@/lib/workspace-onboarding-goals";
+import type { WorkspaceOnboardingChannel, WorkspaceOperatingCountry } from "@/lib/types";
 import type { WizardOnboardingStepId } from "@/lib/messaging-onboarding/wizard-steps";
 
 export { WORKSPACE_ONBOARDING_GOAL_VALUES };
