+++ app/lib/workspace-onboarding-goals.ts
@@ -1,0 +1,13 @@
+/**
+ * Product goals chosen during workspace onboarding.
+ * Shared by client UI and server (non-`.server.ts`).
+ */
+
+export const WORKSPACE_ONBOARDING_GOAL_VALUES = [
+  "live_call",
+  "ivr",
+  "sms_blast",
+] as const;
+
+export type WorkspaceOnboardingGoal =
+  (typeof WORKSPACE_ONBOARDING_GOAL_VALUES)[number];
